Fix bug for original update
You can run this shell again to update v2ray or kcptun.

+++ install.sh
@@ -135,7 +135,9 @@ handle_file()
     fi
     unzip v2ray.zip "*/v2ray" && rm v2ray.zip && mv v2ray* v2ray
     chmod 755 v2ray/v2ray
-    echo 请修改自行本配置文件 > v2ray/config.json
+    if [ ! -f v2ray/config ]; then
+        echo 请修改自行本配置文件 > v2ray/config.json
+    fi
 }
 
 check_CommandLineTools
